requirement.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jwcrypto==1.3.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
zope.interface==5.4.0
